+++ torchao/quantization/subclass.py
@@ -30,8 +30,6 @@ __all__ = [
     "Int8DynamicallyQuantizedLinearWeight",
     "Int8WeightOnlyQuantizedLinearWeight",
     "Int4WeightOnlyQuantizedLinearWeight",
-    "LinearActQuantizedTensor",
-    "to_linear_act_quantized",
 ]
 
 
@@ -599,156 +597,3 @@ class Int4WeightOnlyQuantizedLinearWeight(QuantizedLinearWeightBase):
         )
         int_data = aten._convert_weight_to_int4pack(input_int4x8, inner_k_tiles)
         return int_data, scales_and_zeros, False, groupsize, inner_k_tiles
-
-
-class LinearActQuantizedTensor(torch.Tensor):
-    """
-    Applies activation quantization for linear operator
-    """
-    def __new__(
-        cls,
-        original_weight_tensor: torch.Tensor,
-        input_quant_func: Callable,
-    ):
-        kwargs = {}
-        dtype = original_weight_tensor.dtype
-        kwargs["dtype"] = dtype
-        kwargs["requires_grad"] = False
-        kwargs["device"] = original_weight_tensor.device
-        shape = original_weight_tensor.shape
-        return torch.Tensor._make_wrapper_subclass(cls, shape, **kwargs)  # type: ignore[attr-defined]
-
-    def __init__(
-        self,
-        original_weight_tensor: torch.Tensor,
-        input_quant_func: Callable,
-    ):
-        self.original_weight_tensor = original_weight_tensor
-        self.input_quant_func = input_quant_func
-
-    def __tensor_flatten__(self):
-        return ["original_weight_tensor"], [self.input_quant_func]
-
-    @classmethod
-    def __tensor_unflatten__(
-        cls, tensor_data_dict, tensor_attributes, outer_size, outer_stride
-    ):
-        original_weight_tensor = tensor_data_dict["original_weight_tensor"]
-        input_quant_func, = tensor_attributes
-        return cls(
-            original_weight_tensor,
-            input_quant_func,
-        )
-
-    @classmethod
-    def from_float(cls, input_float, input_quant_func):
-        return cls(input_float, input_quant_func)
-
-    @classmethod
-    def __torch_function__(cls, func, types, args=(), kwargs=None):
-        kwargs = {} if kwargs is None else kwargs
-
-        if func is torch.nn.functional.linear:
-            input_tensor, weight_tensor, bias = (
-                args[0],
-                args[1],
-                args[2] if len(args) > 2 else None,
-            )
-            if isinstance(weight_tensor, LinearActQuantizedTensor):
-                input_quant_func = weight_tensor.input_quant_func
-                original_weight_tensor = weight_tensor.original_weight_tensor
-                aqt = input_quant_func(input_tensor)
-                return torch.nn.functional.linear(aqt, original_weight_tensor, bias)
-
-        with torch._C.DisableTorchFunctionSubclass():
-            return func(*args, **kwargs)
-
-    def _apply_fn_to_data(self, fn):
-        return self.__class__(
-            fn(self.original_weight_tensor),
-            self.input_quant_func,
-        )
-
-    def _get_to_kwargs(self, *args, **kwargs):
-        device, dtype, _, memory_format = torch._C._nn._parse_to(*args, **kwargs)
-        device = self.device if device is None else device
-        dtype = self.dtype if dtype is None else dtype
-        memory_format = (
-            memory_format if memory_format is not None else torch.preserve_format
-        )
-        kwargs = {
-            "device": device,
-            "dtype": dtype,
-            "memory_format": memory_format,
-        }
-        return kwargs
-
-    def to(self, *args, **kwargs):
-        kwargs = self._get_to_kwargs(*args, **kwargs)
-        return self.__class__(
-            self.original_weight_tensor.to(**kwargs),
-            self.input_quant_func,
-        )
-
-    def __torch_dispatch__(cls, func, types, args, kwargs):
-        if (
-            func in [aten.mm.default, aten.addmm.default]
-            and args[0].is_floating_point()
-        ):
-            if func == aten.addmm.default:
-                assert args[1].shape[-1] == args[2].shape[0], (
-                    f"need mat1 shape: {args[1].shape} final"
-                    f"dim to match mat2 shape: {args[2].shape} first dim "
-                )
-                input_tensor, weight_tensor, bias = (
-                    args[1],
-                    args[2],
-                    args[0],
-                )
-                input_quant_func = weight_tensor.input_quant_func
-                original_weight_tensor = weight_tensor.original_weight_tensor
-                aqt = input_quant_func(input_tensor)
-                return func(bias, aqt, original_weight_tensor)
-            else:
-                # aten.mm.default
-                assert args[0].shape[-1] == args[1].shape[0], (
-                    f"need mat1 shape: {args[0].shape} final dim"
-                    f"to match mat2 shape: {args[1].shape} first dim"
-                )
-                input_tensor, weight_tensor = (
-                    args[0],
-                    args[1],
-                )
-                input_quant_func = weight_tensor.input_quant_func
-                original_weight_tensor = weight_tensor.original_weight_tensor
-                aqt = input_quant_func(input_tensor)
-                return func(aqt, original_weight_tensor)
-
-        if func is aten.detach.default:
-            return return_and_correct_aliasing(
-                func, args, kwargs, args[0]._apply_fn_to_data(torch.detach)
-            )
-
-        if func is aten.clone.default:
-            return return_and_correct_aliasing(
-                func, args, kwargs, args[0]._apply_fn_to_data(torch.clone)
-            )
-
-        if func is aten._to_copy.default:
-            return return_and_correct_aliasing(
-                func,
-                args,
-                kwargs,
-                args[0].to(*args[1:], **kwargs)._apply_fn_to_data(torch.clone),
-            )
-
-        if func is aten.t.default:
-            return return_and_correct_aliasing(
-                func, args, kwargs, args[0]._apply_fn_to_data(torch.t)
-            )
-
-        raise NotImplementedError(
-            f"LinearActQuantizedTensor dispatch: attempting to run {func}, this is not supported"
-        )
-
-to_linear_act_quantized = LinearActQuantizedTensor.from_float
